# KaniniAcademy Backend/KaniniAcademy/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "abhi",
    "version": "v1"
  },
  "paths": {
...
    "/api/Certificate/{id}": {
      "delete": {
        "tags": [
          "Certificate"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Certificate"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Certificate"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Certificate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Certificate"
              }
            }
          }
        }
      }
    },
...
    "/api/Course/{id}": {
      "delete": {
        "tags": [
          "Course"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Course"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Course"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Course"
              }
            }
          }
        }
      }
    },
    "/api/CourseResource": {
      "get": {
        "tags": [
          "CourseResource"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CourseResource"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CourseResource"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CourseResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CourseResource/{id}": {
      "delete": {
        "tags": [
          "CourseResource"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CourseResource"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CourseResource"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CourseResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CourseResource"
              }
            }
          }
        }
      }
    },
    "/api/Creators": {
      "get": {
        "tags": [
          "Creators"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Creator"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Creators"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Creator"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Creators/{id}": {
      "delete": {
        "tags": [
          "Creators"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Creator"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Creators"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Creator"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Creator"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Certificate": {
      "type": "object",
      "properties": {
        "cid": {
          "format": "int32",
          "type": "integer"
        },
        "cname": {
          "format": "int32",
          "type": "integer"
        },
        "duration": {
          "type": "string"
        },
        "projects": {
          "type": "string"
        },
        "videos": {
          "format": "byte",
          "type": "string"
        },
        "image": {
          "format": "byte",
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "keySkills": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "howItWorks": {
          "type": "string"
        },
        "syllabus": {
          "type": "string"
        },
        "faqQues": {
          "type": "string"
        },
        "faqAns": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Course": {
      "type": "object",
      "properties": {
        "courseId": {
          "format": "int32",
          "type": "integer"
        },
        "courseName": {
          "type": "string"
        },
        "courseProvider": {
          "type": "string"
        },
        "image": {
          "format": "byte",
          "type": "string"
        },
        "rating": {
          "format": "int32",
          "type": "integer"
        },
        "createdBy": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "CourseResource": {
      "type": "object",
      "properties": {
        "courseId": {
          "format": "int32",
          "type": "integer"
        },
        "video": {
          "format": "byte",
          "type": "string"
        },
        "thumbnail": {
          "format": "byte",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "review": {
          "type": "string"
        },
        "resources": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Creator": {
      "type": "object",
      "properties": {
        "creatorId": {
          "format": "int32",
          "type": "integer"
        },
        "creatorName": {
          "type": "string"
        },
        "creatorImage": {
          "format": "byte",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "creatorJob": {
          "type": "string"
        },
        "followers": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}